Update post preview for home tab
