+++ lib/screens/boards/posts/preview/post_preview.dart
@@ -4,7 +4,7 @@ import 'package:guam_community_client/commons/custom_divider.dart';
 import 'package:guam_community_client/models/boards/post.dart';
 import 'package:guam_community_client/providers/posts/posts.dart';
 import 'package:guam_community_client/screens/boards/posts/detail/post_detail.dart';
-import 'package:guam_community_client/screens/boards/posts/preview/post_preview_banner.dart';
+import 'package:guam_community_client/screens/boards/posts/preview/post_preview_interest.dart';
 import 'package:guam_community_client/screens/boards/posts/post_info.dart';
 import 'package:guam_community_client/styles/colors.dart';
 import 'package:guam_community_client/styles/fonts.dart';
@@ -18,8 +18,6 @@ class PostPreview extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    print(context.read<HomeProvider>().idx);
-
     return Padding(
       padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
       child: Container(
@@ -42,26 +40,22 @@ class PostPreview extends StatelessWidget {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              PostPreviewBanner(post),
+              // PostPreviewBanner(post),
               Padding(
-                padding: EdgeInsets.symmetric(vertical: 8),
-                child: CustomDivider(color: GuamColorFamily.grayscaleGray7),
-              ),
-              Padding(
-                padding: EdgeInsets.only(bottom: 8.0),
+                padding: EdgeInsets.only(top: 4.0),
                 child: Row(
                   children: [
                     if (post.pictures.length > 0)
-                    IconButton(
-                      padding: EdgeInsets.only(right: 4),
-                      constraints: BoxConstraints(),
-                      icon: SvgPicture.asset(
-                        'assets/icons/picture.svg',
-                        color: GuamColorFamily.grayscaleGray5,
-                        width: 20,
-                        height: 20,
+                      IconButton(
+                        padding: EdgeInsets.only(right: 4),
+                        constraints: BoxConstraints(),
+                        icon: SvgPicture.asset(
+                          'assets/icons/picture.svg',
+                          color: GuamColorFamily.grayscaleGray5,
+                          width: 20,
+                          height: 20,
+                        ),
                       ),
-                    ),
                     Text(
                       post.title,
                       style: TextStyle(fontSize: 14),
@@ -71,16 +65,46 @@ class PostPreview extends StatelessWidget {
                   ],
                 ),
               ),
-              Text(
-                post.content,
-                maxLines: 4,
-                overflow: TextOverflow.ellipsis,
-                style: TextStyle(
-                  height: 1.6,
-                  fontSize: 13,
-                  fontFamily: GuamFontFamily.SpoqaHanSansNeoRegular,
-                  color: GuamColorFamily.grayscaleGray3,
+              Padding(
+                padding: EdgeInsets.symmetric(vertical: 8),
+                child: CustomDivider(color: GuamColorFamily.grayscaleGray7),
+              ),
+              PostPreviewInterest(post),
+              Padding(
+                padding: EdgeInsets.symmetric(vertical: 4),
+                child: Text(
+                  post.content,
+                  maxLines: 4,
+                  overflow: TextOverflow.ellipsis,
+                  style: TextStyle(
+                    height: 20.8/13,
+                    fontSize: 13,
+                    fontFamily: GuamFontFamily.SpoqaHanSansNeoRegular,
+                    color: GuamColorFamily.grayscaleGray3,
+                  ),
                 ),
+              ),
+              Row(
+                children: [
+                  Text(
+                    post.boardType,
+                    style: TextStyle(
+                      fontSize: 12,
+                      height: 19.2/12,
+                      fontFamily: GuamFontFamily.SpoqaHanSansNeoRegular,
+                      color: GuamColorFamily.grayscaleGray4,
+                    ),
+                  ),
+                  Spacer(),
+                  Text(
+                    (DateTime.now().minute - post.createdAt.minute).toString() + "분 전",
+                    style: TextStyle(
+                      fontSize: 10,
+                      fontFamily: GuamFontFamily.SpoqaHanSansNeoRegular,
+                      color: GuamColorFamily.grayscaleGray4,
+                    ),
+                  )
+                ],
               ),
               PostInfo(
                 post: post,
